pyproject.toml
[tool.poetry]
name = "juin24cmlops_rakuten"
version = "0.1.0"
description = "A starting point for building an end-to-end machine learning pipeline."
authors = [
    "Jeremy Rava <jeremy.rava@outlook.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/DataScientest-Studio/juin24cmlops_rakuten_2"
repository = "https://github.com/DataScientest-Studio/juin24cmlops_rakuten_2"
packages = [{include = "juin24cmlops_rakuten_2"}]

[tool.poetry.dependencies]
python = "^3.10.14"
keras = "^3.6.0"
tensorflow = "^2.17.0"
pandas = "^2.2.3"
pymongo = ">=4.9.0,<4.10.0"
mlflow = "^2.17.0"
fastapi = "^0.115.3"
typing-extensions = "^4.12.2"
nltk = "^3.9.1"
bs4 = "^0.0.2"
python-multipart = "^0.0.12"
scikit-learn = "^1.5.2"
pillow = "^11.0.0"
motor = "^3.6.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
tqdm = "^4.65.0"
pytest = "^8.3.3"
uvloop = "^0.21.0"
aiohttp = "^3.10.10"
tenacity = "^9.0.0"

[tool.poetry.group.test.dependencies]
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
src = "src"

[tool.pytest.ini_options]
pythonpath = ["."]